Presented hg_repo and git_repo as fixtures.
Closes #36
