+++ newsfragments/36.feature.rst
@@ -1,0 +1,1 @@
+Presented hg_repo and git_repo as fixtures.
