+++ pyproject.toml
@@ -59,3 +59,6 @@ doc = [
 ]
 
 [tool.setuptools_scm]
+
+[project.entry-points]
+pytest11 = {"jaraco.vcs.fixtures" = "jaraco.vcs.fixtures"}
